requirements.txt
tweepy==4.14.0
python-dotenv==1.0.0
flask==2.3.3
flake8==6.1.0
black==23.10.0
